Error handeling for project edit CMS

+++ cms/functions.php
@@ -223,6 +223,17 @@ function get_projects_info(){
                 array_push($specs, $spec);
             }
             $row['spec'] = $specs;
+            
+            $row['img'] = array();
+            $path = "img/projects/project" . $row['id']  . ".";
+            for($i=0;$i<PROJECT_IMAGE_NUMBER;$i++){
+                if(is_file($path.$i.".jpg"))
+                    array_push ($row['img'], $path.$i.".jpg");
+                else
+                    array_push ($row['img'], DEFAULT_IMAGE);
+            }
+            $row['draw'] = is_file($path."draw.pdf") ? $path."draw.pdf" : false;
+            $row['thumb'] = is_file($path."thumb.jpg") ? $path."thumb.jpg" : DEFAULT_IMAGE;
             array_push($out, $row);
         }
         return $out;
@@ -241,9 +252,11 @@ function add_project($prj){
         $q = "INSERT INTO projects (id, name)"
                . " VALUES ('" . $prj['id'] . "', '" . $prj['name'] . "')";
         $res = mysqli_query($con,$q);
-        return $res;
+        if(!$res)
+            return "Project ID already exists!";
+        return true;
     }else{
-        return false;
+        return "Cannot connect to database. Try again!";
     }
 }
 
@@ -253,14 +266,28 @@ function add_project($prj){
  */
 function delete_project($id){
     $con = connect_db();
+    $path = "img/projects/project" . $id  . ".";
+    $des = "img/projects/old";
+    for($i=0;$i<6;$i++){
+        if(is_file($path.$i.".jpg")){
+            rename($path.$i.".jpg", $des.".jpg");
+        }
+    }
+    rename($path."thumb.jpg", $des.".jpg");
+    rename($path."draw.pdf", $des.".pdf");
+    
     if($con){
         $q = "DELETE from projects WHERE id= " . $id; 
         $res1 = mysqli_query($con,$q);
+        if(!$res1)
+            return "The project is not deleted!";
         $q = "DELETE from project_spec WHERE project_id= " . $id;
         $res2 = mysqli_query($con,$q);
-        return $res1 && $res2;
+        if(!$res2)
+            return "The spesificatins are not deleted!";
+        return true;
     }else{
-        return false;
+        return "Cannot connect to database. Try again!";
     }
 }
 
@@ -269,38 +296,55 @@ function delete_project($id){
  * 
  */
 function save_file($des, $file){
-    $allowedExts = array("jpg", "jpeg", "png");
-    $allowedType = array("image/jpeg", "image/jpg", "image/pjpeg" , "image/png", "image/x-png");
-    $temp = explode(".", $file["name"]);
-    $path_jpg = str_replace("_", ".", $des).".jpg";
-    $path_pdf = str_replace("_", ".", $des).".pdf";
-    $extension = strtolower ( end($temp) );
-    if (   in_array($file["type"], $allowedType) 
-        && $file["size"] < 2000000
-        && in_array($extension, $allowedExts)
-        && $file["error"] === 0 ) {
-        if (!file_exists($path_jpg)) {
-            file_put_contents($path_jpg, file_get_contents("img/def.jpg"));
-        }
-        move_uploaded_file($file["tmp_name"], $path_jpg);
-        return true;
-    } else if (    in_array($file["type"], array("application/pdf")) 
-                && $file["size"] < 10000000
-                && in_array($extension, array("pdf"))
-                && $file["error"] === 0 ) {
-                if (!file_exists($path_pdf)) {
-                    // var_dump($path);
-                    file_put_contents($path_pdf, file_get_contents("image/def.pdf"));
+    if($file["size"] !== 0){
+//        var_dump($des);
+//        var_dump($file);
+        $allowedExts = array("jpg", "jpeg", "png");
+        $allowedType = array("image/jpeg", "image/jpg", "image/pjpeg" , "image/png", "image/x-png");
+        $temp = explode(".", $file["name"]);
+        $path_jpg = str_replace("_", ".", $des).".jpg";
+        $path_pdf = str_replace("_", ".", $des).".pdf";
+        $extension = strtolower ( end($temp) );
+        if (  in_array($file["type"], $allowedType) && in_array($extension, $allowedExts) ){
+            if($file["size"] < 2000000){
+                if( $file["error"] === 0 ) {
+                    if (!file_exists($path_jpg)) {
+                        file_put_contents($path_jpg, file_get_contents("img/def.jpg"));
+                    }
+                    move_uploaded_file($file["tmp_name"], $path_jpg);
+                    return true;
+                }else{
+                    return "One of images: ". $file["error"];
                 }
-                move_uploaded_file($file["tmp_name"], $path_pdf);
-                return true;
-    } else {
-        return false;
-    }
+            }else{
+               return "Image size should be less than 2M."; 
+            }
+
+        } else if ( in_array($file["type"], array("application/pdf") ) && in_array($extension, array("pdf"))  ){
+            if($file["size"] < 1000000){
+                if( $file["error"] === 0 ) {
+                    if (!file_exists($path_pdf)) {
+                        // var_dump($path);
+                        file_put_contents($path_pdf, file_get_contents("image/def.pdf"));
+                    }
+                    move_uploaded_file($file["tmp_name"], $path_pdf);
+                    return true;
+                }else{
+                    return "PFD: ". $file["error"];
+                }                
+            }else{
+               return "PDF size should be less than 10M."; 
+            }
+        } else {
+            return ".jpg, .png, .pdf ONLY!";
+        }
+    }else{
+        return true;
+    }    
 }
 
 /*
- * remove the files and edit the project information.
+ * removes the selected files and edits the project information.
  * 
  */
 function edit_project($data){
@@ -309,9 +353,11 @@ function edit_project($data){
         foreach($data["remove"] as $key => $value){
             $ext = ".jpg";
             $path = "img/projects/project" . $project_id . "." . $key;
+            $des = "img/projects/old";
             if($key === "draw")
                 $ext = ".pdf";
-            rename($path.$ext, $path.".old".$ext);
+            if(!rename($path.$ext, $des.$ext))
+                return $path.$ext." is not removed!";
         }
     }
     $con = connect_db();
@@ -319,22 +365,25 @@ function edit_project($data){
         if(isset($data["spec"])){
             $q = "DELETE FROM project_spec WHERE project_id= " . $project_id;
             $res1 = mysqli_query($con,$q);
+            if (!$res1)
+                return "Specifications are not removed!";
             foreach($data["spec"] as $spec){
                 if(trim($spec['name']) != '' && trim($spec['value']) != ''){
                     $q = "  INSERT INTO project_spec (project_id, name, value)
                             VALUES ('" . $project_id . "', '" . $spec['name'] ."', '" . $spec['value'] . "')";
                     $res2 = mysqli_query($con,$q);
+                    if(!$res2)
+                        return "Specifications are not added!";
                 }
             }
         }
-        //var_dump($data);
         $q = "UPDATE snzeng.projects"
            . " SET name= '" . $data['name'] . "', type= '" . $data['type'] . "', year= '" . $data['year'] . "', snippet= '" . $data['snippet'] . "', description= '" . $data['description'] . "', address= '" . $data['address'] . "', projects.order= '" . $data['order'] . "'"
            . " WHERE id= " . $project_id ;
-        $res3 = mysqli_query($con,$q);
-        return $res1 && $res2 && $res3;
+        mysqli_query($con,$q);
+        return true;
     }else{
-        return false;
+        return "Cannot connect to database. Try again!";
     }
     
 }
